requirements.txt
gradio==4.44.1
pandas
chromadb
langchain
langchain-google-genai
langchain-chroma
deep-translator
python-dotenv
/